Handy scripts to bump CSS and JavaScript versions
These scripts simply change the URL of the stylesheet/script by
passing an ignored GET parameter to them, like ?version=4. The server
ignores the `version` parameter and simply returns the file, but the
client thinks that it is a new URL and therefore hits the server even
when it has an old version cached. This forces the same effect as
evicting that file from the cache on the browser.