Disable SIMBAD DSO popups on Steve Gottlieb's notes because they slow the site too much
Longer term solution is to optimize this so that things happen very
snappily. So instead of creating the entire ShadowDOM elements in
connectedCallback, we can do that on the mouse hover
instead. Moreover, we can store the CSS in a stylesheet rather than
having that string written out in JavaScript and simply link it into
the ShadowDOM.